clippy fixes

+++ lib/kv-router/src/concurrent_radix_tree.rs
@@ -1139,7 +1139,7 @@ mod tests {
 
             for i in 0..10 {
                 indexer
-                    .apply_event(create_store_event(0, i, vec![i as u64], None))
+                    .apply_event(create_store_event(0, i, vec![i], None))
                     .await;
             }
 
@@ -1159,7 +1159,7 @@ mod tests {
                     .apply_event(create_store_event(
                         worker_id,
                         1,
-                        vec![1, 2, worker_id as u64 + 10],
+                        vec![1, 2, worker_id + 10],
                         None,
                     ))
                     .await;
